Add execution_diag callback.  It dumps callback info verbosely.
This is a WIP.  Things have to be in their own directories for a few
reasons.  v2 introduces a number of capabilities and issues, and while
it's possible to have combined plugins, these are v1 OR v2 for now.

https://docs.ansible.com/ansible/porting_guide_2.0.html#callback-plugins

In v2 it also appears that the whitelist mechanism is no longer needed.
Plugins are loaded by their presence in the dir ref'd in callback_plugins
or in the ":" delimited list in ANSIBLE_CALLBACK_PLUGINS.

Note: that any CallbackModule class in those locations *including subdirs*
      are loaded.  e.g. ANSIBLE_CALLBACK_PLUGINS=plugins/v2_callback will
      load both profile_timeline and execution_diag.